requirements.txt
alembic==1.4.3
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysqlclient @ file:///D:/Masai%20School/mysqlclient-1.4.6-cp38-cp38-win32.whl
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
